Restore original lcl_ScAddress_Parse_OOo() behaviour detecting external
It's quite unclear why exactly this would even be necessary, but

    commit 296baa2fb6dd4150a7855114093a9703cdc18b09
    CommitDate: Tue Oct 5 18:14:37 2010 -0400

        Ported calc-extref-interpreter-rework-*.diff from ooo-build.

introduced a heuristic whether a reference could be a so far
unrecognized external reference and used String.Erase(n) to strip
a detected possible filename extension starting from the '.' dot
to form a sheet name.

    commit f02b52ca0d76afaf08c8ec3524ae76b0b57e4aaa
    CommitDate: Tue Oct 8 10:03:08 2013 +0200

        convert sc/inc/address.hxx from String to OUString

then changed the String.Erase(n) to OUString::replaceAt(n, 1, "")
which of course does not the same but just removes the '.' dot
character. In that commit this was the only replacement doing
that, others already used a correct OUString::copy().

Just restore the original intention, whyever..

Change-Id: Icc46115f9f25eaf1ed727c233fb1c69942c4b1af
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124238
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins