tdf#124904 sw: don't show deletions in referenced text
In fields, strikethrough/underline formatting of
track changes is not available. If the referenced
text of the fields contained partial deletions,
the field showed them as normal text, e.g. mixing up
the text of a heading with its removed typos.

Now these deleted parts are removed, except in the
following case: if the referenced text was completely
removed, show the whole deleted text in the reference
field, adding strikethrough to its tooltip text to help
the review process.

Note: this allows to handle both use cases, i.e. when
only a small part of a heading was modified (e.g. removing
typos), and 2) removing whole chapters with headings,
where the fields in the deleted text are more readable
with the text of the deleted headings, showing the
original text content.

See also commit 04204377b870402a3169c770508151b34e0d3055
"tdf#61154 sw Table of Contents: remove tracked deletion".

Change-Id: I98bd8ded1591ae3c84957998e65f3cccac2a7884
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124843
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>