ScriptForge - (SF_Calc) First/Last * Row/Column/Cell accept sheets and ranges
The methods
   FirstCell
   FirstColumn
   FirstRow
   LastCell
   LastColumn
   LastRow
   SheetName
accept uniformly either a sheet name or a range name.
They return a single cell as an absolute address (First/LastCell),
a String (SheetName) or a Long.

The range name is to be understood as the extended notation for ranges.

FirstCell/Column/Row accept sheet names. However they will then always return
either the "A1" cell or the value 1.

The LastCell/Column/Row return the real last xxx for ranges, and
the last USED xxx when the argument is a sheet.

The A1Style() method surrounds from now on sheet names
with single quotes only when necessary.

The GetRangeAddress() method has been dropped as redundant with the actual
methods. (Was new in 7.3 => no effect on upward compatibiity)

Change-Id: Iff1ce437ee27dfb8481361f648b3f7a33ae2c4b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124904
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>