qt (>=5.14): Don't create native window for QWidget in QtFrame::screen
For Qt >= 5.14, directly use 'QWidget::screen' (added in Qt 5.14)
to retrieve the screen that the widget is on, rather than first
retrieving a window handle, which forces a platform-native
window to be created.

Due to QTBUG-75766 [1], forcing the creation of
native windows also leads to mouseMoveEvents
not being reliably emitted by Qt, s.

    commit 2a90cbaf121393c45b9d9a00f7f93d05156e6e01
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat May 11 21:31:33 2019 +0200

        tdf#122293 qt5: Use "alien widgets" by default on Wayland

for more details.

The behaviour described in tdf#122293 started showing
up again after

    commit d3c531b9e310d63a689a7601d33d86173a1c4fbb
    Date:   Sun Oct 31 02:33:46 2021 +0200

        tdf#145363 Qt reparent modal dialogs on show

but that just seems to have been a side-effect of
'QWidget::screen' now being called in different ways
than previously.

[1] https://bugreports.qt.io/browse/QTBUG-75766

Change-Id: Ic11923cef5e704c09494f96b19157372a869ae40
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124997
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>