lok: don't pretty print html for online
followup for "lok: don't limit line length in HTMLWriter"

Pretty printing makes html more readable for human but introduces
lots of new line marks which are later interpreted in the client
as a valid new lines in the document content.

This was causing multiplying the line breaks when inserting hyperlinks
where we read current selection's content and then we pass it back
to the core as link's content (with added new lines).

This change needs cypress change to not contain "\n" at the beginning
of some content read from the document.

This also afects copying from the online document to clipboard.

Change-Id: I2b17d62398d947fcf1d3fb1ed6005c3063d114f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136893
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Henry Castro <hcastro@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141479
Tested-by: Jenkins
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>