PPTX: implement native handling of <a:clrScheme> children
This was already handled by converting them to raw colors at import
time. This commit imports the color scheme contents (the 12 colors) into
the doc model.

This is a dependency to export them back to PPTX and to be able to
update these colors on the UI by picking a different theme.

Change-Id: I177de4f15d5f0e628669998d1cda7db24220b2eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125651
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins