avoid only painting to windows in LOK mode, not invalidations
5bd1d371b84e356e5ade3cfa2981fd0c39e2f0ad disabled painting
to windows in LOK mode, because it doesn't make sense as
those windows are not even visible, but it tried to do it
by avoiding even invalidations. That can trigger problems
because some dialogs actually are forwarded to clients in LOK
mode, and apparently they need the invalidations to proceed
normally (the Format->Cell dialog in Online didn't repaint
scrolling of the Date listbox, I don't know what exactly
the actual problem was, but mnPaintFlags were different
than when not avoiding invalidations).

So revert most of the relevant commits and keep only avoiding
actual drawing, which is what 90e756184bb10c6140fff9590532b57f4
did.

Change-Id: I4e212966830f4c690e6143b9d5fedbc24619e1ed
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125674
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>