tdf#131150 sw menu: use nSlotId=0 to avoid adding to set
This fixes a regression in LO 6.4 caused by
commit 5f6e172d13c59c2ec184e10f6d96ea7c63766f2a

Not surprisingly, adding a new condition that multiple
other items match on caused a regression.

bFlag is always set to false at the top of the loop.
In the above commit, bFlag was not set to true,
and so thus it would always be false.
So just set the SlotId to 0
to avoid rSet.Put - like so many other cases already do.

Change-Id: I02e9ca9d8dd75db519e801ad0e86fc8d541ae20a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125688
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <samuel.mehrbrodt@allotropia.de>
Reviewed-by: Justin Luth <jluth@mail.com>