ucb: webdav-curl: avoid pointless HEAD requests for DAV resources
Content::getPropertyValues() has a funny conditional which was changed
in commit aaaca83d234ae32a1faa747dae0b74e826852f8b "Resolves: #i121922#
Fix wrong condition".

But the new condition also appears wrong, there is no need to do any
further requests if all properties have already been retrieved, so the
check for m_bDidGetOrHead could just be omitted.

There is a surprising result in the Remote Files dialog when connecting
to a Sharepoint 16 server: all properties have been retrieved from the
directory given as Root, but because of !m_bDidGetOrHead the branch is
taken anyway.

The HEAD request results in a "HTTP/1.1 302 Redirect" to
$ROOT/Forms/AllItems.aspx, HEAD on this URL gives a "HTTP/1.1 200 OK"
and then subsequently a PROPFIND on that very same URL results in
"HTTP/1.1 404 NOT FOUND".

Physics claims that a single observation should be sufficient to
determine whether an URL exists or not, but Sharepoint is apparently
able to maintain this quantum superposition indefinitely, hence the
Remote Files dialog doesn't display anything as the 404 causes an exception.

Try to fix this by requiring a missing property again before initiating
the HEAD request for DAV resources.

Change-Id: I1239762948b6abd1f8fc097edd4a16cb6b75ca7a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125826
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>