tdf#145873 Fix FILEOPEN: EMF file not displayed
EMF can have WMF embedded inside it. The sample PPTX file contains an
EMF file consisting of several GDIcomment records in which embed a
Placeable WMF file without some records related to the size of the
WMF. It reaches pStm->SetError( SVSTREAM_FILEFORMAT_ERROR ); in
emfio/source/reader/wmfreader.cxx:1982 which causes to the problem
in loading the file.

With this fix, the EMF file will be displayed. Without the fix, the
EMF file will be displayed as blank.

Regression from
    commit e7ea0b73ab5183b487fce003f40dd2483a5e0744
    tdf#88163 Fix font size for placeable wmf files

The fix can be tested with:

    make CPPUNIT_TEST_NAME="testTdf145873" -sr \
        CppunitTest_sd_svg_export_tests

Change-Id: I04babd5029ba123fdd03758eef2be69faf45fda8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125796
Tested-by: Jenkins
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>