tdf#144244: fix crash when FILESAVE a calc file after delete some columns
see bt here:
https://bugs.documentfoundation.org/attachment.cgi?id=176507

Regression from 971a4fc75e1e218afdf4550d470a802d9f47df4f
tdf138138 Apply shape shift also to NoRotatedAnchor
The 'normal' anchor (as on screen) is connected to the snap rectangle,
the NoRotatedAnchor is connected to the logic rectangle. They differ,
if the shape is transformed, e.g. rotated. Error was, that values of
the 'normal' anchor were applied to NoRotatedAnchor instead of
calculating the shift of NoRotatedAnchor independently. The error
becomes only visible on save, because there the NoRotatedAnchor is
used. Effected shape types are legacy shapes, text boxes and
transformable OLEs.
I have not tested, whether this fix would work for LO 7.0 too.

Let's apply the same controls as we do with ScDrawObjData retrieved with GetObjDataTab

Change-Id: I73cb3dd3c47a7ec55667e498d4a62842ce63d861
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125852
Tested-by: Jenkins
Tested-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-by: Eike Rathke <erack@redhat.com>