tdf#145873 Fix FILEOPEN: EMF file not displayed #2
The previous patch cfc81805125c0c1576729893946271968de161c4 only
partially fixed the problem, and the size of the graphical objects
was wrongly calculated. This patch fixes this problem.

The fix can be tested with:

    make CPPUNIT_TEST_NAME="testTdf145873" -sr \
         CppunitTest_vcl_pdfexport

This test exports the PPTX to PDF and checks the size of the objects.
The previous test for this issue which exported to SVG is removed, as
it could not test the current situation.

Change-Id: I97bed99811eaf8236ded84ffe0c1a7d3d1b5bad5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125924
Tested-by: Jenkins
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>