new ScTypedStrData: typically missed argument in CTOR, tdf#142910 follow-up
Pretty much any attempted use of eType was completely wrong and lost.

Regression from

    commit 6b0f0ac92a896dbe83c6f9fced883f02ae6f15e8
    CommitDate: Wed Jul 7 17:44:46 2021 +0200

        tdf#142910 sc filter: fix "greater than" or "smaller than" etc

Most calls to this are missing the "rounded number" argument,
so the enumator is actually accepted as the double fRVal,
and the StringValue eType was left as the default value (Standard),
instead of the intended enumerator.

0.0 looks too much like 0, 0 to even notice in
casual code reading.

This had rendered the type mostly irrelevant.

Change-Id: If4fa69d4b3077981244a2c3a785f80b77f9f9501
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126453
Tested-by: Eike Rathke <erack@redhat.com>
Reviewed-by: Eike Rathke <erack@redhat.com>