tdf#121699 sc: wrap LOKit stuff with isActive
This is a partial revert of LO 6.1
commit 87d07adfb3535d4e03d08538cba60b9bd2a47846.

It is one thing to change code that you don't really
know why it is setup that way, but it is another thing
to ignore the resulting regression reports for 3 years.
Not impressed.

It look me only 1 hour to work through this,
and this is completely outside of my normal
line of programming.

Change-Id: I8671677f38cc1332c6c5a76847801e2680097048
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126500
Tested-by: Jenkins
Reviewed-by: Henry Castro <hcastro@collabora.com>
Reviewed-by: Justin Luth <jluth@mail.com>