tdf#144680 mailmerge toolbar: enable icons unless no resultset
This fixes a problem introduced in LO 6.3 (backported to 6.1.5)
via commit a9fb586aea9f899ac027dc8f8c91bcbda8d9c2a9
tdf#121606: displaying Mail Merge toolbar shouldn't activate connection

Before 6.1, a user needed to run the mailmerge wizard
(or start the toolbar manually) in order to have
the mailmerge toolbar become available. In either of those cases,
the connection to the database was attempted and thus enabled
or disabled the various buttons based on the status. All good.

In 6.1, the toolbar was started on document load,
as soon as database fields were detected.
Sounds reasonable - good context sensitive UI.

The problem is that databases can often be password protected,
or on slow, remote sites. Thus the 6.3 commit mentioned above,
which disabled most buttons until the connection was established.

Well, having disabled buttons isn't too helpful,
and basically put us back to needing the extra step of
running the mail-merge wizard anyway.
The only problem with leaving them enabled is that we need to
ensure that we don't run the functions with bad results.

So sure - this means that a user thinks they can use the
mail-merge, attempt it, and then may find the buttons disable
when it won't work. That is the same as the forward/backwards
buttons though, so no big deal. Plus it is good feedback that
something isn't right (which will rarely ever be true).

The only time the buttons disable is for non-existant resultsets.
[Note that buttons do NOT disable for a wrong password,
or for a remote connection that can't be established.
And that is what we want anyway, so that further attempts
can be made after fixing the underlying problem.]

Change-Id: I3e82d2d3b35d04632650933c768cb0b5c006cb6f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126625
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>